Rename component to UpdateIndividualWorkPlaceCard for consistency
